# application/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "A comprehensive API for managing a mechanic shop including customers, mechanics, service tickets, and inventory"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "JWT Authorization header using the Bearer scheme. Example: 'Bearer {token}'"
paths:
...
  /mechanics:
    post:
      tags:
        - Mechanics
      summary: "Create a new mechanic"
      description: "Add a new mechanic to the system"
      parameters:
        - in: "body"
          name: "body"
          description: "Mechanic information"
          required: true
          schema:
            $ref: "#/definitions/CreateMechanicPayload"
      responses:
        201:
          description: "Mechanic created successfully"
          schema:
            $ref: "#/definitions/MechanicResponse"
          examples:
            application/json:
              id: 1
              first_name: "Mike"
              last_name: "Smith"
              email: "mike.smith@mechanicshop.com"
              phone: "555-9876"
              specialty: "Engine Repair"
              hourly_rate: 85.50
        400:
          description: "Invalid input"
          schema:
            $ref: "#/definitions/ErrorResponse"
        409:
          description: "Mechanic with this email already exists"
          schema:
            $ref: "#/definitions/ErrorResponse"

    get:
      tags:
        - Mechanics
      summary: "Returns all Mechanics"
      description: "Endpoint to retrieve a list of all mechanics."
      responses:
        200:
          description: "Retrieved Mechanics Successfully"
          schema:
            $ref: "#/definitions/AllMechanics"
  /mechanics/{mechanic_id}:
    get:
      tags:
        - Mechanics
      summary: "Get a specific mechanic by ID"
      description: "Retrieve detailed information about a specific mechanic"
      parameters:
        - in: "path"
          name: "mechanic_id"
          type: "integer"
          required: true
          description: "ID of the mechanic to retrieve"
      responses:
        200:
          description: "Mechanic retrieved successfully"
          schema:
            $ref: "#/definitions/MechanicResponse"
        404:
          description: "Mechanic not found"
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      tags:
        - Mechanics
      summary: "Update a mechanic"
      description: "Update an existing mechanic's information"
      parameters:
        - in: "path"
          name: "mechanic_id"
          type: "integer"
          required: true
          description: "ID of the mechanic to update"
        - in: "body"
          name: "body"
          description: "Updated mechanic information"
          required: true
          schema:
            $ref: "#/definitions/UpdateMechanicPayload"
      responses:
        200:
          description: "Mechanic updated successfully"
          schema:
            $ref: "#/definitions/MechanicResponse"
        400:
          description: "Invalid input"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Mechanic not found"
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      tags:
        - Mechanics
      summary: "Delete a mechanic"
      description: "Remove a mechanic from the system"
      parameters:
        - in: "path"
          name: "mechanic_id"
          type: "integer"
          required: true
          description: "ID of the mechanic to delete"
      responses:
        200:
          description: "Mechanic deleted successfully"
          schema:
            $ref: "#/definitions/DeleteResponse"
        404:
          description: "Mechanic not found"
          schema:
            $ref: "#/definitions/ErrorResponse"

  # Service Ticket Endpoints
...
definitions:
...
  CreateMechanicPayload:
    type: "object"
    properties:
      first_name:
        type: "string"
        example: "Mike"
      last_name:
        type: "string"
        example: "Smith"
      email:
        type: "string"
        format: "email"
        example: "mike.smith@mechanicshop.com"
      phone:
        type: "string"
        example: "555-9876"
      specialty:
        type: "string"
        example: "Engine Repair"
      hourly_rate:
        type: "number"
        format: "float"
        example: 85.50
      hire_date:
        type: "string"
        format: "date"
        example: "2024-01-15"
    required:
      - first_name
      - last_name
      - email
  UpdateMechanicPayload:
    type: "object"
    properties:
      first_name:
        type: "string"
        example: "Mike"
      last_name:
        type: "string"
        example: "Johnson"
      email:
        type: "string"
        format: "email"
        example: "mike.johnson@mechanicshop.com"
      phone:
        type: "string"
        example: "555-4321"
      specialty:
        type: "string"
        example: "Transmission Repair"
      hourly_rate:
        type: "number"
        format: "float"
        example: 95.00
    required:
      - first_name
      - last_name
      - email
  MechanicResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        example: 1
      first_name:
        type: "string"
        example: "Mike"
      last_name:
        type: "string"
        example: "Smith"
      email:
        type: "string"
        example: "mike.smith@mechanicshop.com"
      phone:
        type: "string"
        example: "555-9876"
      specialty:
        type: "string"
        example: "Engine Repair"
      hourly_rate:
        type: "number"
        example: 85.50
  AllMechanics:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
          example: 1
        first_name:
          type: "string"
          example: "Mike"
        last_name:
          type: "string"
          example: "Smith"
        email:
          type: "string"
          example: "mike.smith@mechanicshop.com"
        phone:
          type: "string"
          example: "555-9876"
        specialty:
          type: "string"
          example: "Engine Repair"
        hourly_rate:
          type: "number"
          example: 85.50

  # Service Ticket Definitions
...
  DeleteResponse:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Successfully deleted"
  ErrorResponse:
    type: "object"
    properties:
      error:
        type: "string"
        example: "An error occurred"
      errors:
        type: "object"